requirements.txt
absl-py==0.9.0
asn1crypto==1.3.0
astor==0.8.0
blinker==1.4
cachetools==3.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.1.1
cryptography==2.8
cssselect2==0.3.0
decorator==4.4.2
defusedxml==0.6.0
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.4
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
rdp==0.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
svgpathtools==1.3.3
svgwrite==1.3.1
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-probability==0.8.0rc0
termcolor==1.1.0
tinycss2==1.0.2
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.12.1
